switch to use the same variable for margin as our grid + added more description and a js TODO
I tried using our grid but it added more complexity both in regard to classes in HTML and in CSS for the browser to interpret. Using the same variable as our grid ensures that changes to our grid will be reflected in f-hero margin, while leaving out the complexity of our grid which in this case, is not needed.